Add contentReference elements, more ValueSet examples, insert rules with concept context (#133)
* Add Element Rules may have content reference

The section for Add Element Rules is updated with information about
defining an element using a content reference. The grammar is updated to
the current version present in SUSHI, which includes content reference
elements. Some other new portions of the grammar will be documented in
other change sets.

* Add complex examples for Value Set definitions

Value Set rules can be combined to include a system, value sets, and
filters. The new examples illustrate some of these combinations.

* Add example of inserting rule set with concept context

* Mk feedback (#134)

* Suggestions to update Mint's PR

* Update input/pagecontent/reference.md

Co-authored-by: Mint Thompson <mathompson@mitre.org>

* Update input/pagecontent/reference.md

Co-authored-by: Mint Thompson <mathompson@mitre.org>

* Add missing space

Co-authored-by: Mint Thompson <mathompson@mitre.org>

Co-authored-by: markkramerus <mkramer@mitre.org>