Add missing `Id` keyword for Mappings
In some places of the Mapping description, the Id keyword is missing. It was already present in the examples and in the keyword descriptions, but missing from the keyword table and the textual description of mapping keywords. The keyword is supported by SUSHI and not required, so I've added it as suggested (since StructureDefinition.mapping.identity is needed anyway).
The quick reference sheet is still wrong.